requirements.txt
feedparser==6.0.10
python-dotenv==1.0.0
requests==2.31.0
slack-sdk==3.21.3
openai==1.3.5
praw==7.7.1